requirements.txt
-e Store
-e Webserver
-e Config
-e Database
Pillow==2.6.1
PyYAML==3.11
bcrypt==1.0.2
gnureadline==6.3.3
lxml==3.4.0
mongo==0.2.0
motor==0.3.3
pycparser==2.10
readline==6.2.4.1
tornado==4.0.2
